+++ include/shared_types/Control.h
@@ -1,4 +1,4 @@
-#pragma once
+﻿#pragma once
 #include <QString>
 
 namespace shared_types {
@@ -15,11 +15,18 @@ class Control {
     bool isAuto() const;
     void setIsAuto(bool newIsAuto);
 
+    int getMoisturePercentage() const;
+    void setMoisturePercentage(int newMoisturePercentage);
+    float getTemperature() const;
+    void setTemperature(float newTemperature);
+
   private:
-    bool isRaining_      = false;
-    bool isHeating_      = false;
-    int lightPercentage_ = 0;
-    int windPercentage_  = 0;
-    bool isAuto_         = false;
+    bool isRaining_         = false;
+    bool isHeating_         = false;
+    int lightPercentage_    = 0;
+    int windPercentage_     = 0;
+    bool isAuto_            = false;
+    int moisturePercentage_ = 0.0f;
+    float temperature_      = 0.0f;
 };
 } // namespace shared_types